Updating logic

+++ ws/src/rpc/call.rs
@@ -1,7 +1,8 @@
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
+/// Call: [<MessageTypeId>, "<MessageId>", "<Action>", {<Payload>}]
 pub struct Call {
-    pub message_type_id: i64,
+    pub message_type_id: usize,
     pub message_id: String,
     pub action: String,
     pub payload: serde_json::Value,
@@ -9,7 +10,7 @@ pub struct Call {
 
 impl Call {
     pub fn new(
-        message_type_id: i64,
+        message_type_id: usize,
         message_id: String,
         action: String,
         payload: serde_json::Value,
