+++ ws/src/rpc/call_error.rs
@@ -1,7 +1,8 @@
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone)]
 #[serde(rename_all = "camelCase")]
+/// CallError: [<MessageTypeId>, "<MessageId>", "<errorCode>", "<errorDescription>", {<errorDetails>}]
 pub struct CallError {
-    pub message_type_id: i64,
+    pub message_type_id: usize,
     pub message_id: String,
     pub error_code: String,
     pub error_description: String,
@@ -10,7 +11,7 @@ pub struct CallError {
 
 impl CallError {
     pub fn new(
-        message_type_id: i64,
+        message_type_id: usize,
         message_id: String,
         error_code: String,
         error_description: String,
